Reorganize core modules.
