+++ Cargo.toml
@@ -7,8 +7,7 @@ edition = "2018"
 [workspace]
 members = [
   ".",
-  "./kekbit_core",
-  
+  "kekbit_core",
 ]
 
 [dependencies.kekbit_core]
