+++ kekbit_core/src/lib.rs
@@ -1,7 +1,6 @@
+pub mod api;
 mod header;
-pub mod reader;
 pub mod shm;
 pub mod tick;
 mod utils;
 mod version;
-pub mod writer;
