+++ src/lib.rs
@@ -1,6 +1,8 @@
 pub mod core {
-    pub use kekbit_core::reader::*;
-    pub use kekbit_core::shm::*;
+    pub use kekbit_core::api::*;
+    pub use kekbit_core::shm::reader::*;
+    pub use kekbit_core::shm::shm_reader;
+    pub use kekbit_core::shm::shm_writer;
+    pub use kekbit_core::shm::writer::*;
     pub use kekbit_core::tick::*;
-    pub use kekbit_core::writer::*;
 }
